+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Added new options for usage sections: `demoData`, `highlight` and `highlightProps`
+
 ## 1.0.0-beta.69 (29-09-2022)
 
 - Implemented `tooltip` as a view wide option